requirements.txt
pyserial
pycaw
comtypes